Fix container tests

+++ model/service/service_test.go
@@ -8,8 +8,7 @@ import (
 func TestService(t *testing.T) {
 	s := New()
 
-	// TODO(brutella): int64 cast should not be required!
-	if is, want := s.GetID(), int64(model.InvalidID); is != want {
+	if is, want := s.GetID(), model.InvalidID; is != want {
 		t.Fatalf("is=%v want=%v", is, want)
 	}
 	if is, want := len(s.GetCharacteristics()), 0; is != want {
